add upload&reportnig

+++ resources/views/manageUser.blade.php
@@ -4,6 +4,7 @@
     <tr>
         <td><a href="/manage">Manage Artikel</a></td>
         <td><a href="/manage/user">Manage User</a></td>
+        <td><a href="/printuser" target="_blank">Cetak PDF</a></td>
     </tr>    
 </div>
 @section('judul','Manage User')
@@ -12,7 +13,11 @@
 <div class="tambah mb-2">
     <a href="/user/add" class="btn btn-primary">Tambah User</a>
 </div>
-
+    @if(session('status'))
+    <div class="alert alert-success">
+        {{session('status')}}
+    </div>
+    @endif
 <table class="table table-bordered table-striped ">
     <thead class="thead-dark">
         <tr>
@@ -28,12 +33,12 @@
         @foreach ($User as $u)
         <tr>
             <td>{{$loop->iteration}}</td>
-            <td>{{$u->imageurl}}</td>
+            <td><img width="50px" width="50px"src="{{asset('storage/'.$u->image) }}" alt=""></td>
             <td>{{$u->name}}</td>
             <td>{{$u->email}}</td>
             <td>{{$u->updated_at}}</td>
             <td><a href="/user/edit/{{$u->id}}" class="badge badge-warning">Edit</a>
-                <a href="/user/delete/{{$u->id}}" class="badge badge-danger">Delete</a></td>
+                <a href="/user/delete/{{$u->id}}" class="badge badge-danger" onclick="return confirm('Yakin ingin menghapus User')">Delete</a></td>
         </tr>
         @endforeach
     </tbody>
